+++ files/private-chef-cookbooks/private-chef/recipes/nginx.rb
@@ -31,8 +31,8 @@ nginx_addon_dir = File.join(nginx_etc_dir, "addon.d")
   nginx_tempfile_dir
 ].each do |dir_name|
   directory dir_name do
-    owner owner
-    group group
+    owner OmnibusHelper.new(node).ownership['owner']
+    group OmnibusHelper.new(node).ownership['group']
     mode node['private_chef']['service_dir_perms']
     recursive true
   end
@@ -94,8 +94,8 @@ remote_directory nginx_html_dir do
   files_owner "root"
   files_group "root"
   files_mode "0644"
-  owner owner
-  group group
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
   mode node['private_chef']['service_dir_perms']
 end