+++ files/private-chef-cookbooks/private-chef/recipes/postgresql.rb
@@ -51,7 +51,8 @@ end
 ####
 
 directory postgresql_log_dir do
-  owner node['private_chef']['user']['username']
+  owner OmnibusHelper.new(node).ownership['owner']
+  group OmnibusHelper.new(node).ownership['group']
   recursive true
   mode node['private_chef']['service_dir_perms']
 end