+++ files/private-chef-cookbooks/private-chef/recipes/redis_lb.rb
@@ -29,8 +29,8 @@ redis_log_dir = redis['log_directory']
   redis_log_dir,
 ].each do |dir_name|
   directory dir_name do
-    owner owner
-    group group
+    owner OmnibusHelper.new(node).ownership['owner']
+    group OmnibusHelper.new(node).ownership['group']
     mode node['private_chef']['service_dir_perms']
     recursive true
   end