+++ CHANGELOG.md
@@ -28,6 +28,7 @@
 * Fixed the default value for Postgres effective_cache_size
 * Adjust perms to 0750 for all service's log dir
 * Add and use new perms attribute
+* Add an OmnibusHelper method to provide an owner and group hash
 
 ### chef-server-ctl
 * Partition server start/stop in upgrade process